+++ docs/images/badges/accomplishment/power-liberator.png
